refactor: formatTime utils 폴더로 이동
