+++ src/utils/format.js
@@ -1,0 +1,6 @@
+export const formatTime = (unFormattedTime) => {
+  const date = new Date(unFormattedTime)
+  const day = date.toLocaleDateString()
+  const time = date.toTimeString().slice(0, 5)
+  return `${day} ${time}`
+}
